refactor: unify inserter
