+++ src/frontend/src/lib.rs
@@ -28,3 +28,4 @@ mod server;
 pub mod service_config;
 pub mod statement;
 pub mod table;
+pub(crate) mod inserter;
